Refactor alerter
