+++ consensus/src/alerts/io.rs
@@ -1,17 +1,17 @@
 use super::*;
 
-pub struct IO<'a, H: Hasher, D: Data, MK: MultiKeychain> {
+pub struct IO<H: Hasher, D: Data, MK: MultiKeychain> {
     pub messages_for_network: Sender<(NetworkMessage<H, D, MK>, Recipient)>,
     pub messages_from_network: Receiver<NetworkMessage<H, D, MK>>,
     pub notifications_for_units: Sender<ForkingNotification<H, D, MK::Signature>>,
     pub alerts_from_units: Receiver<Alert<H, D, MK::Signature>>,
-    pub rmc: ReliableMulticast<'a, H::Hash, MK>,
+    pub rmc: ReliableMulticast<H::Hash, MK>,
     pub messages_from_rmc: Receiver<RmcMessage<H::Hash, MK::Signature, MK::PartialMultisignature>>,
     pub messages_for_rmc: Sender<RmcMessage<H::Hash, MK::Signature, MK::PartialMultisignature>>,
     pub alerter_index: NodeIndex,
 }
 
-impl<'a, H: Hasher, D: Data, MK: MultiKeychain> IO<'a, H, D, MK> {
+impl<H: Hasher, D: Data, MK: MultiKeychain> IO<H, D, MK> {
     pub fn rmc_message_to_network(
         &mut self,
         message: RmcMessage<H::Hash, MK::Signature, MK::PartialMultisignature>,
